Remove nargs="+" from --nodes and --roles expect "comma, string"
Currently --nodes and --roles have nargs="+" which is one or
more. Furthermore these are made into a list by argparse [1].
Passed as a list into the ansible playbook invocation will
not work.

Instead use defaults, remove nargs so one argument, which is a
string (the default) and which can be comma delimited if you want
to specify more than one --nodes "one, two" or --roles "foo, bar"
This will be passed to ansible-playbook verbatim in [2] and this
is the supported format "comma, delimited, string" for multiple
values.

[1] https://docs.python.org/3/library/argparse.html#nargs
[2] https://github.com/openstack/tripleo-common/blob/3ac0ae3ec2fa583f5ae52c580ff127bff35bb08a/tripleo_common/actions/ansible.py#L387

Change-Id: Ic8a9ed228e59845b904a7f839ce2da98ece72b66