Use container_images_file for all image prepare
The recently merged change Ie8543bed3363aac539cfd35e04e8dfaaa166ebc3
added five new options to undercloud.conf to drive the container
prepare during undercloud install. However this is not nearly as
flexible as the prepare workflow for overcloud deploy, which can have
multiple entries in the ContainerImagePrepare parameter list.

This change removes those new options, and uses the specified
container_images_file file as the source for the ContainerImagePrepare
parameter *or* the container image parameters.

If no file is specified, the environment has a default
ContainerImagePrepare parameter set which will result in a working
undercloud based on the defaults in kolla_builder.

Change-Id: Ic1e6eb5ff0aaeaded279c97139adc9fdebb12857
Blueprint: containerized-undercloud