Change default output-dir to be $HOME
In https://review.openstack.org/#/c/542875 we created --output-dir with
the default to $HOME/.undercloud-heat-installer which is not backward
compatible when you upgrade from a non-containerized undercloud.

Indeed, it'll fail to re-use the same passwords as the previous
deployment because the undercloud-passwords.conf file is in $HOME.

For backward compatibility, we should keep $HOME as default output_dir,
and allow deployers to override it.

Change-Id: I3b8464a3a2761d6cef319f2a1d08baedc9d8cc3e