add lower-constraints job
Create a tox environment for running the unit tests against the lower
bounds of the dependencies.

Create a lower-constraints.txt to be used to enforce the lower bounds
in those tests.

Add openstack-tox-lower-constraints job to the zuul configuration.

See http://lists.openstack.org/pipermail/openstack-dev/2018-March/128352.html
for more details.

Update the requirement for tripleo-common to reflect reality, and update
the constraint for tenaciy.

Change-Id: Ifdf8dd1c782d6d6ac575d6fd5364dcb083e805a9
Signed-off-by: Doug Hellmann <doug@doughellmann.com>