Add upgrade-[prepare|converge].yaml into upgrade cli
This explicitly adds the upgrade-prepare.yaml environment file
or the upgrade-converge.yaml environment file, into the stored
swift plan, before continuing with the previous workflow (here
heat stack update in both cases). See dependency below in tht
and this tie-in is the biggest downside I can see.

The idea is to stop doing it in the tripleo-common like [1][2]
and move all the upgrade|update|ffwd-upgrade cli to use the
same and include the expected environment file.

Co-Authored-By: Lukas Bezdicka <social@v3.sk>
Co-Authored-By: Jiri Stransky <jistr@redhat.com>
[1] https://github.com/openstack/tripleo-common/blob/6090d32b516f92837fc8b01b7a55dc97ed3e8a77/tripleo_common/actions/package_update.py#L62-L76
[2] https://github.com/openstack/tripleo-common/blob/6090d32b516f92837fc8b01b7a55dc97ed3e8a77/tripleo_common/actions/plan.py#L492-L502
Depends-On: Icfe494e3219d6d6cd3251f75bb4329fc4d793c3c
Change-Id: I1288fe68ae8af02a5d77390d237ec467d88e43d2