Correct stack and add ceph_ansible_playbook in upgrade unittests.
The stack name being mocked was not matching
with the one used in the args. This didn't fail
because at the time of asserting mock_upgrade
we were passing still the wrong stack name, not
the one from the arguments.

This patch also adds the ceph_ansible_playbook
arg to verifylist.

Change-Id: I854ce652995c3381c389d5f4d7e6b68d881f78bb