Enable cleanup by default for undercloud install
The goal of this change is to remove the randomly generated data that we
we use during the installation of the undercloud by default.  This
change switches the cleanup configuration in the undercloud.conf to
be true by default so we cleanup the generated heat temporary files by
default. The undercloud deploy flag still remains the same where it is
not currently enabled by default if you leverage the 'openstack
undercloud deploy' command.

Change-Id: I027a76d2f37a5af5bd0a599783287632298acf3f
Related-Bug: #1761810