Implement LocalContainerRegistry parameter
LocalContainerRegistry is a new parameter that is set to local_ip by
default. It'll be used in TripleO Heat Templates when deploying a local
container registry, to bind the service on this address.

Before, it was done in instack-undercloud and automatically binded on
local-ip. Now, it'll be a parameter and still defaulted to the local-ip
used on the undercloud.

It's part of the efforts to containerize the undercloud and deploy
Docker with Ansible.

Change-Id: Iab2615f9be9e107d85d509f75bbfee01fe04b88a