Implement TripleoImagePrepare command
Running:

  $ openstack tripleo container image prepare default \
      --output-env-file prepare-default.yaml \
      --local-push-destination
  $ openstack tripleo container image prepare -e prepare-default.yaml

Will read a Heat template with ContainerImagePrepare, prepare containers
and upload to a registry if needed.

The idea is to replace the other commands used by the overcloud for any
use case: undercloud, overcloud or any cloud.

One of the goals here is to execute this process before starting the
containers while deploying OpenStack on any cloud.

Change-Id: Ie4b7951147f5a1aec654982e21296a749fdd865c
Blueprint: container-prepare-workflow