Capture external process output
This change implements a function that can be used to run external
command and log their out. It also switches out the execvp that we used
for running ansible for a subprocess call so that we can properly
capture the output from the ansible run. Additionally because
the execvp never returned, we were unable to do any cleanup actions
after ansible ran. This change switches out the output to use the logger
rather than print statements so that in the future we can properly
leverage the logger for writing out installation logs.

Change-Id: Ibb04145905c0103f271f85282df161cc43e6c9ee
Related-Blueprint: containerized-undercloud