Make usage of --*-flavor and --*-count an error
The flavor and count CLI params have been deprecated since Newton.
Obviously, we'd like to remove them, but these params are rather
crucial to earlier deployments. As such, we should have these be an
error for a cycle (or two) to give deployers a chance to change their
command lines and get the environment files set up properly.

Depends-On: https://review.openstack.org/#/c/648564/
Change-Id: I94376c5f25c354a9fcdbca11548f7e3eedbd9491
Related-Bug: #1768244