Stop trying to converge during Ceph upgrade command
This had no effect as we only updated the plan and not the
stack. We'll do a normal converge after Ceph upgrade command, which
will also take care of putting back the normal deploy
CephAnsiblePlaybook.

Furthermore it was prepending the converge env file to a list of envs
which already had the *prepare* env file at a latter position (we
reused the env list from the prepare call :) ), so it kept doing
prepare operation instead of converge.

And perhaps most importantly, the messages printed to user are
misleading since no converge was really happening.

Also adds support for --container-registry-file to ceph-upgrade command,
while not necessary, some of the functions called by ceph-upgrade
command expect to receive the container-registry parameters.

Co-Authored-By: Giulio Fidente <gfidente@redhat.com>
Change-Id: I025eac40f8bda5f23c789e7fef1a9e9b49947f66
Partial-Bug: #1768586