Create generic deployment action
Previously we've created an undercloud deploy action for use with the
containerized undercloud. In order to support other deployments that may
not be underclouds, we should rename the interface to something more
generic. This change creates a tripleo deploy action and switches the
undercloud deployment mechanism to leverage the tripleo deployment
action.

The tripleo deployment action has a new parameter called --standalone to
indicate if we want to trigger an standalone deploy scenario. The
undercloud deployment is a special standalone deployment so the existing
undercloud deploy command passes standalone as true to the tripleo
deployment action. Eventually we should roll in the existing overcloud
deployment into this tripleo deploy command to align the two interfaces.

Change-Id: I377b304ad78f54701971c7a29f71d612c62f5dc0
Related-Blueprint: all-in-one