Add --stack-only cli arg
--stack-only can also be used to disable the config-download workflow,
and has the same effect as --no-config-download, but with a better name.
The deprecation warning is removed as technically it's not deprecated to
deploy the Heat stack and not use the config-download workflow.

What is deprecated is using the
environments/disable-config-download-environment.yaml environment from
tripleo-heat-templates, and there is a deprecation warning in that file.

Change-Id: I78e9aa9cf4906e7b4e502f781a1ec0d94dfaebde