Reorganize the config options for deployment
In order to support a basic configuration file for a standalone
deployment, this change moves the configuration items out into their own
classes for inheriting configuration items since the undercloud install
is basically a special form of a standalone. By creating a standalone
configuration file, we will be able to use enable_* for services to
allow for a quick deployment of specific architectures.

Change-Id: Id07acaa2dfb8432d97dabccbb256a206cbf94e1f
Related-Blueprint: all-in-one