Convert enable-ssh-admin.sh to python
Instead of using the script from the templates, use python for the
enable-ssh-admin logic. This will allow for more properly handling
failures.

This also fixes a race condition where sshd has not already started on
some of the nodes before we try and connect via ssh. A timeout is added
where we wait for the port to come up. If the timeout has passed and the
port is still not up, then an exception is raised.

Change-Id: I3431d2ec724a880baf0de8f586490d145bedf870
Closes-Bug: #1769230