Implement 'openstack tripleo upgrade'
Adding a new command:
$ openstack tripleo upgrade

This new command will run underneath:
$ openstack tripleo deploy --upgrade

The only difference when running this command is that we'll execute
upgrade tasks generated by config-download before the deployment tasks.

It'll be used by the containerized undercloud to upgrade by re-using the
upgrade tasks in THT, that were designed for the overcloud. We'll try to
re-use them so we avoid code duplication.

Change-Id: Ic932b3b308c83bc72fff4c0ef8ce94e878f33997