Prepare call, add --modify-role and --modify-vars
This allows the prepare command to set modify_role and modify_vars in
the call to container_images_prepare. The result will be an images
file that has extra options set in every entry.

The actual modification on the images is done during the upload
command, driven by these extra options.

Change-Id: Ia1dc7a1b7f777ac3e000f0c2e155be6330c05a2c
Blueprint: container-prepare-workflow
Depends-On: I2c877a96264b351b4fc8527a3e40b87ddcb4f9a5