Add .deployment.v1.deploy_on_servers to ffwd-upgrade prepare
Before we update the heat stack for the ffwd-upgrade prepare we
can use this to run any required commands on the overcloud.

The current use case is removal of the
/usr/libexec/os-apply-config/templates/etc/os-net-config/config.json
Otherwise after the upgrade prepare is executed, overcloud nodes
will have empty config.json for os-net-config (also discussed [0])

The current implementation of .deployment.v1.deploy_on_servers assumes
all servers are managed by nova. This is OK for N..Q ffu but we will
need to revisit in future, see [1] for more info.

[0]: https://bugzilla.redhat.com/show_bug.cgi?id=1561255
[1]: https://bugs.launchpad.net/tripleo/+bug/1770356
Related-Bug: 1758161
Related:
    tripleo-common Ic138c4925c5c96d1cf718af7ae59bcf5f0ba96dc
Change-Id: I77fdc9deab97c785725f09b8529c512950753f86