Update HostnameMap generation
The HostnameMap is used in combination with the
deployed-neutron-port.yaml for the Containerized undercloud. Previously
we were hardcoding the value assuming 'undercloud' in the hostname from
the stack and role. Because these can be customized, we should generate
the hostname map value from the stackname being deployed and the primary
role from the roles_data.

Change-Id: I503dcc831b78d1c271f166ffd6d97fdc2cbd618f
Related-Blueprint: all-in-one