Make standalone role name configurable
Previously the default role name for the standalone deployment was
'Undercloud'. This was not configurable so unless you used a new role
named 'Undercloud' with 'openstack tripleo deploy', it wouldn't deploy.
This change adds a new command line parameter named --standalone-role to
the command.  The undercloud deploy has been updated to continue to pass
'Undercloud' to this parameter so the containerized undercloud remains
the same.

Change-Id: I4637465d2a65b2b45633eb82f9ef576a450d6265
Related-Blueprint: all-in-one