Move undercloud.conf to the undercloud config namespace
To match the generation for standalone.conf, this change moves the
config generation out to the undercloud config namespace instead of
including it in the undercloud_config namespace from the actions.
Eventually we'll want to handle the configuration parsing elsewhere so
we need to move this logic out so it can be shared and consumed via the
tripleo_deploy action.

Change-Id: I6d8bd04d2547a513dfe46d5144283e45366efeb3
Related-Blueprint: all-in-one