DeployIdentifier should be provided for any parameter actions
DeployIdentifier is update during the deployment workflow in the
tripleo_common, but all the parameter update and get workflows,
before the depoloyment will fail since the parameter is not provided
for the heat stack validation. Adding a dummy entry for the DeployIdentifier
in the tripleoclient which will be replaced to a timestamp in the
deploy workflow.

Closes-Bug: #1759326
Change-Id: I26cc11895a376788d8aad41727676bec73a15d69
(cherry picked from commit 4cb7a225297e804fb00c037904142dea9262475b)