Deployed Server - portmap should match neutron API
The cidr property for a Neutron subnet should contain
the network cidr, not the network prefixlen.

The example portmap for the fake neutron port in tripleo
depleoyed-server is not correct. We should pass the
netowork cidr in the subnets so that the fake neutron
port subnets attribute return something that match
what an actual neutron port would return.

Change-Id: I0f8cad690a178e7b70044c38b2277a51ae3734eb
Related-Bug: #1772396