Mark upgrade_cleanup option as experimental
The feature doesn't work now, the upgrade_tasks that take care of the
cleanup are being refactored by the Upgrade team. The current status of
this feature is that the upgrade will fail in some of the tasks, and
also the tasks aren't idempotent (for the ones which worked).

Until we stabilize it, let's mark it as experimental.

Related-Bug: #1774219
Change-Id: I38c411cae75c419af18a654c741a014502ecd292