Copy user provided files into working tht root
Implements a standard pattern to normalize user provided templates by
copying them into the working tht path.  When a user provides a -e for a
file that already exists within the templates path, it is updated to the
expected working tht root.  For files that are outside of the templates
path, they will be copied into the root of the working tht root and
processed using the process-templates.py during the deployment. This
allows for a standard relative path and supports j2 files.

Note that this means you will not be able to provide two identically
named files that live outside of the templates folder since they will
conflict. An explicit exception is thrown if the system identifies an
already existing file.

Change-Id: I2138a5dcf6a2aff04f7e522cbdf2d5ce94283dc1
Related-Blueprint: all-in-one