tripleo deploy accept plan environment
This enables passing a different plan environment file, and also
avoids hard-coded references to some of the environment files
such as the overcloud-resource-registry-puppet.yaml which we may
want to rename to enable plans per application e.g openstack,
openshift, ceph standalone etc in future.

This also aligns better with the overcloud deploy command.

Change-Id: I61763f62c5d44f098ad60f9426871caef16cd6de