Remove deploy playbook from `update run`
We need the deploy steps as part of update playbook, to execute them
with `serial: 1` (we need to start services on node A before
proceeding to update node B). This is already happening correctly
within the update playbook.

After the update playbook we have been additionaly running deploy
playbook on all nodes in parallel. That seems like a superfluous
operation, in some environments considerably prolonging the time for
minor update. The full deployment will be once again executed on all
nodes in parallel during `update converge`, we don't need to do it in
`update run` too.

Change-Id: I513f655707c8a31f3621b96a756be9c468e3bd8b
Closes-Bug: #1776206