Track fake stack updates for standalone/UC deploy
Puppet relies on stack_action UPDATE for some cases.
Track the ephemeral heat stacks fake state for tripleo
undercloud/standalone deployments to match the puppet
expectations. For such deployments, the heat stack
state is a fake (virtual), as we never
update but always create a new ephemeral heat stack.

When the deployment is finished w/o errors, create the mark
file (unique to the stack name) that is used to indicate to
puppet the stack_action has been changed from CREATE to
UPDATE. The indication is done via a drop-in file with
top level override containing either 'StackAction: CREATE'
or 'UPDATE'. The drop-in is created in the working/templates
directory and named <Stack_name>-stack-vstate-dropin.yaml.

When the deployment fails, remove the mark file, so
the serial re-deployments will be considered as creating
a stack. The --force-stack-update flag keeps the mark file
instead, so the serial re-deployment will be
considered (virtually) updating the heat stack.

For the --output-only mode, only warn users to control
the stack virtual state manually. The state is considered
CREATE, unless there is --force-stack-update specified.

For --dry-run, also log the expected stack virtual
state/action the deployment would go with.

Closes-bug: #1778505

Change-Id: I55dc83acb2ed5ee07b4cf57e25135e6201589ac4
Signed-off-by: Bogdan Dobrelya <bdobreli@redhat.com>