Update undercloud_nameserver help text
Since the undercloud was containerized we set the dns servers
on the neutron ctlplane subnets when installing the undercloud.
(Previously this was a post install step ...) Since dnsservers
set as undercloud_nameservers are configured as nameserver on
the ctlplane subnet(s). The overcloud nodes will use the
dnsservers defined here (Atleast initially, before OsNetConfig).

With change I5f33e06ca3f4b13cc355e02156edd9d8a1f773cd the
requirement to specify DnsServers in the overcloud environment
is removed. It uses get_attr to read the nameservers to use for
overcloud nodes from the ctlplane subnet.

Closes-Bug: #1824347
Change-Id: I6e330479fa78163252f9319b8f90d6d819ce881b