Quick'n'dirty way to at least protect and correct rights on files
Currently the rights on some files are wrong due to wild "sudo" calls
in the deploy process.

The right way would be to get a proper privileges escalation using
oslo.privsep, but as this spec is for Stein, we still need some
corrections for the previous versions.

Please keep in mind this is a quick fix, NOT a final nor proper
solution.

Based on a previously abandonned (unmergeable) proposal.

Change-Id: I418d41c91d008283360ffaaa7b3a38354e405423
Closes-Bug: #1761595
Co-Authored-By: Emilien Macchi <emilien@redhat.com>