Use prefix length for CIDR in DeployedServerPortMap
The cidr key takes the prefix length in DeployedServerPortMap, not the
full CIDR notation. For example, for 192.168.24.1/24, the value should
be 24.

We have to use 'cidr' as the key, even though it is a bit of a misnmoer
because the structure accepted by DeployedServerPortMap has to match
the structure of OS::Neutron::Port resources in Heat.

Change-Id: I05ed17e2a24f9744466adedd86aa4b8852c51fc9