Switch --use-heat to take a boolean and deprecate it
--use-heat parameter is deprecated in Rocky cycle and will be removed in
the future.
When --use-heat / --use-heat=True is set, the undercloud will be
containerized and a warning will be shown for the deprecation.
When --use-heat=False is set, the undercloud won't be containerized.

This allows the case where someone wants to deploy a non-containerized
undercloud even if it's not the recommended way in Rocky.

Co-Authored-By: Alex Schultz <aschultz@redhat.com>
Change-Id: Ibc6fb33d752346d6f8b3a61d74e0bc73c1ac20e3