Always set a hardware architecture on uploaded images
In a multiarch environment if we have an image that doesn't have an
hw_architecture it can be scheduled on any node registered with ironic.
So this means that nova could schedule an x86_64 image on a ppc64le node
(clearly a bad thing).

Avoid that by always setting a hw_architecture image property matching
the kernel of the running system.  This is safe as in a single arch
environment, by definition all nodes are the same ;P and in a multiarch
we're being explict, which may result in NoValidHost exceptions which is
a much clearer error case than scheduling wrong ISA images.

Blueprint: multiarch-support
Change-Id: I082dfe34cf192f76f87e348aecb67e984e0c8ae9