Update output messaging for standalone/undercloud
With the exposing of the tripleo deploy command, the installation and
success messaging should be different in the containerized undercloud
case vs the standalone.  This change adds different messaging to be
added at the end of the undercloud installation and reduces what we
print when standalone is used.

Change-Id: I99ce8abde838f16cabefe641b107e55b508c2cf9
Related-Blueprint: all-in-one