Add --force-stack-create
--force-stack-create will force the stack action to be CREATE when
UPDATE would otherwise be used. This is useful in scenarios when you had
a successful deployment, but did something wrong, so you want to froce a
recreate, for example when using a completely wrong roles file.

--force-stack-create and --force-stack-update are added to an argparse
mutually exclusive group so that both can't be specified at the same
time.

Also adds unit tests around the stack action handling and adds missing
unit tests for --force-stack-update.

Change-Id: Ia3edebfe2271710abc4b84ce090b9e7ddf3c9d88