Don't assume ansible_dir in finally clause
As an exception can occur before ansible_dir is ever defined, we
shouldn't assume it exists in the main finally clause, otherwise an
Undefined exception could be thrown.

Instead, save ansbie_dir as a class variable and default it to None,
that way we can first check if it's been set before trying to use it in
finally.

Change-Id: I90e9718324b2f5e8420b89e8c00b5e366b82e96b