Log tracebacks
Always log the traceback for unexpected exceptions. It's very difficult
to tell where the error came from without the traceback.

It seems this was the intent with
Iad9de7fab0ee740bd20f8facd49f36e6f18d2fb1, but
traceback.format_exception doesn't actually print/log anything, it
returns a list of strings, which are just lost if not actually
saved/logged.

Instead use traceback.print_exc in the main handler to show the
traceback.

Change-Id: If84a41907e1fc782bd5a4608f445e75aa3a5d2fb