Prepare defaults, defer detecting undercloud registry
Now that the undercloud is containerized, the following command will
sometimes need to be run before the undercloud is installed:

  openstack tripleo container image prepare default --local-push-destination

Doing this currently results in push_destination:localhost:8787
because the required network interface doesn't yet exist for the
registry detection logic.

This change will make push_destination be set to True, so that the
registry address is detected at upload time, which is late enough in
the undercloud install to actually work.

Change-Id: I69a30be9372dc2349b3339019be9d45b632f4a74
Closes-Bug: #1783204
Depends-On: I8a35755ac28c419490159f3cb0f60b7d54f3958d