Handle negative/wrong answer in question when upgrading undercloud.
When upgrading the heat-based undercloud a security
question is asked to stop the upgrade if a undercloud
backup was not performed. This patch handles the case
when a negative or wrong answer is provided,
raising a new UndercloudUpradeNotConfirmed exception
which is then captured and displays an informative
log notifying that the upgrade didn't take place.
Also, removing the format call when the undercloud
upgrades fails [0] as the called string message
doesn't need parametrization.

[0] - https://github.com/openstack/python-tripleoclient/blob/03254c84f6c6822b7b87476d4597a19fd2eb5983/tripleoclient/v1/undercloud.py#L158

Change-Id: I80ae52e8a732b2827e8179ce024f92ed52a29394
Closes-Bug: #1783722