New function: run_ansible_playbook
This new function allows to run ansible playbooks
in a convenient way, taking advantage of already existing
run_command_and_log - that latter method has been modified so
that we can get the process full state instead of just the exit
code.

It implements some checks in order to ensure the playbook exists,
and allows to set a temporary ansible configuration file in order
to avoid unwanted behaviors (like retry file creation, callback for
output, and so on).

Change-Id: I5de4d68cc2669d37e7a667209423b9ac842136ff