Drop incorrect reference to heat_launch
I99ce8abde838f16cabefe641b107e55b508c2cf9 updated the failure message to
not actually be a template, however the undercloud was attempting to use
self.heat_launch when formatting the failure message. The heat_launch
code only exists in the TripleoDeploy so this leads to a misleading
error message during failures.

Change-Id: Ica15fe506c06a8557d27675e450646dc96c5e6c4
Closes-Bug: #1783389