Update default value for ntp servers
As we've switched to containerized undercloud, the ntp settings are more
important (and required).  If this is unset, we're actually falling back
to the defauls in the ntp configuration in THT so this patch just
exposes the default correctly in the configuration file.

Change-Id: I98c4d0c3cee2a745e0e5326e209ad91db22ebff1
Closes-Bug: #1793500
(cherry picked from commit 2f02c6503fbfff63f4c120a7cd1c463e5aea7fdd)