Use sync action get_deployment_failures
Instead of using the workflow, which has the extra overhead of opening a
websocket and polling for messages and the workflow result, just use the
mistral action for get_deployment_failures directly. This is much
simpler.

It also fixes a bug where messages from other running workflows using
the same "tripleo" queue were polluting the output of the "overcloud
failures" command.

Change-Id: Ie774a5698515ba7e43dc0755042fb476eb241fc7
Closes-Bug: #1794277