Switch Heat Launcher to use Podman instead of Docker when containerized
When the Heat Launcher is containerized, we don't use Docker anymore but
Podman which will be the default container CLI in Stein cycle.

Note: we didn't maintain Docker support for the launcher as there is no
use case where we need its support for this cycle. The Heat Launcher can
be non-containerized if needed.

Blueprint podman-support
Change-Id: I1acf0f2949390f97f3b9ea3b2a49811dcaf84d98