Update deployment status when deploying or failed
There were a couple spots where the deployment status was not properly
updated. The first was when the stack failed, the deployment status also
needed to be set to failed.

The second is when we start the config_download process, the status
needs to be set to deploying.

Depends-On: I87b97bcab3e8cec6bdbb79c0acb95a375bb2d2fa
Change-Id: Ida519df829450c1ce9216049d76bc098176c273b
Closes-Bug: #1795640
(cherry picked from commit e478e97231ba9fd0fe019a2aacb150ee712fe2bd)