Create DeployIdentifier for standalone
There is a bunch of update logic in the framework that expects that a
DeployIdentifier will be set on updates in order to function correctly.
This change adds the generation of a DeployIdentifier to our update
logic to ensure that we are passing it to the deployment on updates.

Change-Id: Icf9a8f1c8049e46cc33541c5c40fc0d706a5869a
Closes-Bug: #1796421
(cherry picked from commit 5d4cebf3959ff706ce26f364fb905a94b36f58bd)