Fix defaults to templates, roles-file and plan-environment
This change updates the --templates parameter to now have a default
which makes it optional and only necessary when something different than
the default.  Additionally we're removing the defaults as it causes
issues, when someone download the overcloud plan on the undercloud
because it would use the current working director to load the file.
Now it will use the --templates folder as a base for the default value
and will take the specific file if --roles-file or
--plan-environment-file is specified.

Change-Id: Ia0c8a7caaf8f24ef1e5b0858d57827633f2473e9
Closes-Bug: #1797922
Co-Authored-By: Alex Schultz <aschultz@redhat.com>