Undercloud: Disable persisting VIPs in /etc/hosts
This adds the use-dns-for-vips.yaml which disables persisting the VIP
hosts to /etc/hosts. This is because sometimes we might be setting the
CloudName to an IP (for using IPs as keystone endpoints). Without this,
we'll get an erroneous value there.

Conflicts:
	tripleoclient/tests/v1/undercloud/test_undercloud.py
	tripleoclient/v1/undercloud_config.py

Change-Id: Ie28f51d6729f5f4b940bfcc8e77968dd57d41669
Closes-Bug: #1799724
(cherry picked from commit 6789f2e9a90de492f3b6c0b9f954ec0703948208)