Only return plans if available from the message
There appears to be a race condition in the list_deployment_plans such
that if another message is recieved before the plan results, we attempt
to return the plans payload which may not exist. This change adds a
check to ensure the payload contains plans before trying to return the
values.

Change-Id: I58da6f699fe8257e3ddb89cd3516889944c4ad92
Closes-Bug: #1802592
(cherry picked from commit 82ace31be5fcd5e28a8d253cda3b221d8dabf140)