Mock configuration load in unit tests
Previously if you had an undercloud.conf in your home directory, the
tests would load it and use values from it as the defaults for the tests
if they were not overridden in the tests. This could lead to incorrect
failures. This change mocks the config loading out from the install and
upgrade actions for the undercloud to prevent this loading.

Change-Id: I57a12f0151bc65f26a7e880b0f1f367a7711ead9
Closes-Bug: #1803387