undercloud: enable undercloud_config parameter for validations
If TripleO Validations are enabled, we want to export UndercloudConfigFilePath
so the validations can work from the Mistral Executor container.

UndercloudConfigFilePath is set even if the preflight validations runs
are ommitted (when using the --no-validations option).

These parameters will be used in THT to feed hieradata and also bind
mount the undercloud configuration file.

Change-Id: I38f6d562ab452f765672e7cc4e71206b562c471b
Related-Bug: #1805825