Improve error message around hieradata override file
If the file specified is empty, we currently throw a generic message
about it not being in the correct format. Let's add a check to make sure
the file is not empty before attempting to parse it as yaml.

Change-Id: I702da1170484c0d10c3ae836ec3f7550975a341c
Closes-Bug: #1810436
(cherry picked from commit 8ddcc2cd145451335a90b6816c8110c65dc73998)