Use existing stack parameter for DeployIdentifier
In commit 4cb7a225297e804fb00c037904142dea9262475b we changed
to set the DeployIndentifier to ''. This looks like an incorrect
change. It would set the DeployIdentifier to '' (from the existing
unique value), if  --skip_deploy_identifier is used and this would
result in deployments applied on nodes, whereas we want exactly
the opposite. This checks the existing value in the stack for an
update.

Conflicts:
	tripleoclient/tests/v1/overcloud_deploy/test_overcloud_deploy.py

Change-Id: Idd4178aa06aba5668dec6d8ba0427f57dfe5b5b6
Related-Bug: #1759326
(cherry picked from commit d550c19485efa388950288faaecf821a1b210138)