Fail if too many nameservers are provided
The nameservers are limited by neutron to the max configured servers
which is 5 by default. Since we do not offer a way to tune this value,
let's fail the user early if they are providing too many nameservers.
If a user wants to provide more than 5, they can do so by increasing the
max_dns_nameservers in neutron and setting the DnsServers via extra
environment files.

Change-Id: I51ed6f8bb109df7f9038bbf691097d50ce9e54b3
Closes-Bug: 1812919